# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Travel Agency API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: getTrips
      tags: [trips]
      summary: Get all trips
      parameters:
        - name: destination
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips, admin]
      summary: Create a new trip (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}:
    get:
      operationId: getTripById
      tags: [trips]
      summary: Get trip by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTrip
      tags: [trips, admin]
      summary: Update trip (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "200":
          description: Trip updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips, admin]
      summary: Delete trip (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Trip:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        destination:
          type: string
        country:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        duration:
          type: integer
          description: Duration in days
        maxCapacity:
          type: integer
        availableSpots:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        featured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        includes:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - destination
        - country
        - price
        - duration
        - maxCapacity
        - availableSpots
        - startDate
        - endDate
        - featured
        - rating
        - reviewCount
        - includes
        - createdAt
    CreateTripRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        destination:
          type: string
        country:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        duration:
          type: integer
        maxCapacity:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        featured:
          type: boolean
        includes:
          type: array
          items:
            type: string
      required:
        - title
        - description
        - destination
        - country
        - price
        - duration
        - maxCapacity
        - startDate
        - endDate
        - featured
        - includes
...